% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_history.R
\name{embed_history}
\alias{embed_history}
\title{Combine RIS search results with search record}
\usage{
embed_history(ris_input, search_record)
}
\arguments{
\item{ris_input}{The input RIS file to be read into R.}

\item{search_record}{The search record saved as a JSON file in
the standardised search data format.}
}
\value{
An RIS file containing the search record embedded in the
custom field of the first record in the file.
}
\description{
Read an RIS file containing a set of search results
into R and combine with a JSON search history file, embedding the
search history into a custom field within the first record in the
RIS file. The output is then exported as a modified RIS file.
}
\examples{
\dontrun{
embed_history(ris_input = 'references.ris', search_record = 'search_record.JSON')
}
}
