% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_history.R
\name{extract_history}
\alias{extract_history}
\title{Extract search history from modified RIS file}
\usage{
extract_history(ris_file = NULL, save_JSON = FALSE)
}
\arguments{
\item{ris_file}{The input RIS file containing the search history
to be read into R.}

\item{save_JSON}{Logical argument specifying whether to save the
JSON search history file to the working directory. The default is
set to FALSE.}
}
\value{
A formatted text string as a JSON object containing the
search history
}
\description{
Read a modified RIS file containing a set of search
results along with an embedded search history into R and extract
the search history as a JSON formatted object.
}
\examples{
\dontrun{
search_history <- extract_history(ris_file = 'inst/extdata/ris_output.ris', save_JSON = TRUE)
}
}
