% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ris.R
\name{load_ris}
\alias{load_ris}
\title{Read references into R in multiple object formats}
\usage{
load_ris(ris_file = NULL, output, verbose = TRUE)
}
\arguments{
\item{ris_file}{The input RIS file to be read into R.}

\item{output}{The type of R object output specified as one or
more of the following: "list", "dataframe",
"string", "lines". Multiple output types should be listed as a
vector (e.g. 'c("list", "string")').}

\item{verbose}{Logical argument for printing a report to the
console. Default is set to TRUE}
}
\value{
An R object containing the RIS file contents in one or
more of three formats, as specified in 'output'
}
\description{
Read an RIS file into R to produce one or more of
four different formats of R object: a reference list, a data
frame, a text string, or a vector of lines. Multiple formats
facilitate manipulation of the underlying RIS text.
}
\examples{
\dontrun{
refs <- load_ris(output = c("list", "string", "dataframe", "lines"))
}
}
