% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_searchrecord.R
\name{load_searchrecord}
\alias{load_searchrecord}
\title{Read search record into R}
\usage{
load_searchrecord(search_record = NULL, output = "both")
}
\arguments{
\item{search_record}{The input search record saved as a JSON
file.}

\item{output}{Specification of what output format should be
provided: a "list" of items from the search record, a "string"
containing the JSON formatted text (for embedding into an RIS
file), or "both". The default is set to "both".}
}
\value{
A formatted text string as a JSON object containing
the search history
}
\description{
Read a search record file as a JSON object. See
xxx for details of how to format the JSON file.
}
\examples{
\dontrun{
search_history <- load_searchrecord(output = "both")
}
}
